Add unit tests

+++ src/main/java/mx/com/bankaya/assessment/dto/pokeapi/VersionDetailDto.java
@@ -2,10 +2,12 @@ package mx.com.bankaya.assessment.dto.pokeapi;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class VersionDetailDto {
 
-    private Long rarity;
+    private BigDecimal rarity;
     private VersionDto version;
 
 }
